docs: add docs to email functionalities in orders-service
